*: export tidb-server main method
