+++ cli/main_test.go
@@ -11,7 +11,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package main
+package cli
 
 import (
 	"testing"
@@ -27,7 +27,7 @@ var isCoverageServer = "0"
 // and it is used to generate coverage_server.
 func TestRunMain(t *testing.T) {
 	if isCoverageServer == "1" {
-		main()
+		Main()
 	}
 }
